Move exemption presenter methods to a concern for use by multiple presenters
